Create a package for crud operations
