+++ backend/app/crud/base.py
@@ -1,17 +1,7 @@
-from enum import Enum
-
 from pydantic import BaseModel
 from sqlalchemy.orm import Session
 
-from .models import Chat, Document, Message, User
-from .schemas.user import UserCreate
-
-
-class Model(Enum):
-    User = User
-    Chat = Chat
-    Message = Message
-    Document = Document
+from .enums import Model
 
 
 class CRUD:
@@ -49,16 +39,3 @@ class CRUD:
         db.delete(obj)
         db.commit()
         return obj
-
-
-class UserCRUD(CRUD):
-    def __init__(self):
-        super().__init__(Model.User.value)
-
-    def get_user_by_email(self, db: Session, email: str):
-        return db.query(self.model).filter(self.model.email == email).first()
-
-    def create(self, db: Session, obj_in: UserCreate):
-        hashed_password = obj_in.password + "notreallyhashed"
-        del obj_in.password
-        return super().create(db, obj_in, {"hashed_password": hashed_password})
