+++ backend/app/crud/enums.py
@@ -1,0 +1,10 @@
+from enum import Enum
+
+from ..models import Chat, Document, Message, User
+
+
+class Model(Enum):
+    User = User
+    Chat = Chat
+    Message = Message
+    Document = Document
